requirements.txt
tensorflow==2.17.0
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
openpyxl==3.1.5